Added test classes with no implementation

+++ src/main/model/Body.java
@@ -1,8 +1,8 @@
 package model;
 
 public abstract class Body {
-    protected int mass;
-    protected int radius;
+    protected double mass;
+    protected double radius;
 
 
 }
